refactor(observe): share normalizeConfigKeys across grids + test [TH-6080]
The snake_case->camelCase column-config normalizer was a byte-identical inline
copy in TraceGrid, SpanGrid, and Session-grid; reading raw is_visible instead of
the normalized isVisible has shipped blank-table bugs three times (#996, #951,
this PR). Hoist one shared export into LLMTracing/common.js, import it in all
three grids, drop the local copies. Adds a normalizeConfigKeys test pinning
is_visible -> isVisible so the next grid cannot regress it.

Typed column-config contract (the durable fix) tracked in a separate ticket.
